refactor: フォロー関連のファイルをfeaturesディレクトリに集約
